change user id to int in repository funcs

+++ Source/MindKeeper.Api/Data/Repositories/Users/IUserRepository.cs
@@ -6,7 +6,7 @@ namespace MindKeeper.Api.Data.Repositories.Users
 {
     public interface IUserRepository
     {
-        public Task<User> Get(long id);
+        public Task<User> Get(int id);
         public Task<User> Get(string mail);
         public Task<List<User>> GetAll();
         public Task<User> Create(string mail, string passwordHash);
